+++ android/src/main/java/com/webcare/zoom/sdksample/initsdk/InitAuthSDKCallback.java
@@ -1,6 +1,0 @@
-package com.webcare.zoom.sdksample.initsdk;
-
-import us.zoom.sdk.ZoomSDKInitializeListener;
-
-public interface InitAuthSDKCallback extends ZoomSDKInitializeListener {
-}