+++ android/src/main/java/com/webcare/zoom/sdksample/inmeetingfunction/customizedmeetingui/BaseEvent.java
@@ -1,4 +1,0 @@
-package com.webcare.zoom.sdksample.inmeetingfunction.customizedmeetingui;
-
-public interface BaseEvent {
-}