+++ android/src/main/java/com/webcare/zoom/sdksample/inmeetingfunction/zoommeetingui/CustomNewZoomUIActivity.java
@@ -1,6 +1,0 @@
-package com.webcare.zoom.sdksample.inmeetingfunction.zoommeetingui;
-
-import us.zoom.sdk.NewMeetingActivity;
-
-public class CustomNewZoomUIActivity extends NewMeetingActivity {
-}