+++ android/src/main/java/us/zoom/sdksample/MyMeetingActivity.java
@@ -1,0 +1,19 @@
+package us.zoom.sdksample;
+
+import android.os.Bundle;
+import android.view.WindowManager;
+
+import us.zoom.sdk.NewMeetingActivity;
+
+public class MyMeetingActivity extends NewMeetingActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+       
+            getWindow().setFlags(WindowManager.LayoutParams.FLAG_SECURE, WindowManager.LayoutParams.FLAG_SECURE);
+       
+
+    }
+}